#35 refactor: make PEVENT CamelCase
