+++ Ppap/app/src/main/java/com/jeongg/ppap/presentation/component/PDialog.kt
@@ -18,6 +18,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
+import com.jeongg.ppap.presentation.component.util.noRippleClickable
 import com.jeongg.ppap.theme.Dimens
 import com.jeongg.ppap.theme.gray1
