+++ Ppap/app/src/main/java/com/jeongg/ppap/presentation/util/PEvent.kt
@@ -1,8 +1,7 @@
 package com.jeongg.ppap.presentation.util
 
 sealed class PEvent{
-    object SUCCESS: PEvent()
-    object LOADING: PEvent()
-    object NAVIGATE: PEvent()
-    data class TOAST(val message: String): PEvent()
+    object Loading: PEvent()
+    object Navigate: PEvent()
+    data class MakeToast(val message: String): PEvent()
 }
